backend get/delete functions
5 functons working

+++ NeighborlyAPI/deleteAdvertisement/__init__.py
@@ -1,5 +1,6 @@
 import azure.functions as func
 import pymongo
+import os
 from bson.objectid import ObjectId
 
 
@@ -9,14 +10,22 @@ def main(req: func.HttpRequest) -> func.HttpResponse:
 
     if id:
         try:
-            url = "localhost"  # TODO: Update with appropriate MongoDB connection information
+            url = os.environ['myMongoDbconnectionString']
             client = pymongo.MongoClient(url)
-            database = client['azure']
+            database = client['neighborlydB']
             collection = database['advertisements']
-            
+    
+    
             query = {'_id': ObjectId(id)}
-            result = collection.delete_one(query)
-            return func.HttpResponse("")
+            result = collection.delete_one(query) # note: the id is real oid
+            if result is None:
+                query = {'_id':id}
+                result = collection.delete_one(query) # note: the id is just a string in db due to import
+
+            if (result.deleted_count >0):
+                return func.HttpResponse("Successfuly deleted Ad with ID "+id, status_code=200)
+            else:
+                return func.HttpResponse("id not found "+id, status_code=400)
 
         except:
             print("could not connect to mongodb")
